cleanup and more jsdoc
